smart-farming-irrigation-controller/requirements.txt
fastapi
uvicorn
joblib
pydantic
streamlit
requests
pyserial
scikit-learn